+++ Clicked/Locales/enUS.lua
@@ -283,10 +283,10 @@ L["The generated macro for binding '%s' is too long and will not function, pleas
 L["Press any key or ESC to cancel"] = true
 L["%d loaded bindings"] = true
 L["+%d more"] = true
-L["Key Options"] = true
+L["Options"] = true
 L["Enabled"] = true
 L["Disabled"] = true
-L["Enabled by: %s"] = true
+L["May be enabled by: %s"] = true
 L["Group name"] = true
 L["Group icon"] = true
 L["Macro name"] = true